Ajout fonction upload_file et delete_file dans AppController

+++ app/Controller/PartnersController.php
@@ -11,24 +11,14 @@ class PartnersController extends AppController {
 			$d = $this->request->data;
 			$d['Partner']['partner_id'] = null;
 			$d['Partner']['partner_date_added'] = date('Ymd');
-			if ($d['Partner']['partner_img']['error'] == 0) {
-				$tmp_name = $d['Partner']['partner_img']['tmp_name'];
-				$extension = pathinfo($d['Partner']['partner_img']['name'], PATHINFO_EXTENSION);
-				if (in_array(strtolower($extension), array('gif', 'jpeg', 'jpg', 'png'))) {
-					$name = (string)time() . '.' . $extension;
-					if (move_uploaded_file($tmp_name, IMAGES  . '..' . DS . 'files' . DS  . 'images' . DS . 'partners' . DS . $name) == true) {
-						$d['Partner']['partner_img'] = $name;
-						if ($this->Partner->save($d, true, array('partner_id', 'partner_type', 'partner_name', 'partner_desc', 'partner_img', 'partner_to_home', 'partner_to_page', 'partner_date_added'))) {
-							$this->Session->setFlash("Le partenaire à bien été ajouté !", 'notif');
-							$this->redirect(array('controller' => 'partners', 'action' => 'index', 'admin' => true));
-						} else {
-							$this->Session->setFlash("Un problème est survenu !", 'notif', array('type' => 'error'));
-							$this->redirect($this->referer());
-						}
-					} else {
-						$this->Session->setFlash("Un problème est survenu !", 'notif', array('type' => 'error'));
-						$this->redirect($this->referer());
-					}
+			$d['Partner']['partner_img'] = $this->upload_file($d['Partner']['partner_img'], 'partners');
+			if ($d['Partner']['partner_img'] != false) {
+				if ($this->Partner->save($d, true, array('partner_id', 'partner_type', 'partner_name', 'partner_desc', 'partner_img', 'partner_to_home', 'partner_to_page', 'partner_date_added'))) {
+					$this->Session->setFlash("Le partenaire à bien été ajouté !", 'notif');
+					$this->redirect(array('controller' => 'partners', 'action' => 'index', 'admin' => true));
+				} else {
+					$this->Session->setFlash("Un problème est survenu !", 'notif', array('type' => 'error'));
+					$this->redirect($this->referer());
 				}
 			}
 		}
@@ -40,24 +30,15 @@ class PartnersController extends AppController {
 			$d = $this->request->data;
 			$d['Partner']['partner_id'] = $data['Partner']['partner_id'];
 			$this->Partner->id = $data['Partner']['partner_id'];
-			if ($d['Partner']['partner_img']['error'] == 0) {
-				$tmp_name = $d['Partner']['partner_img']['tmp_name'];
-				$extension = pathinfo($d['Partner']['partner_img']['name'], PATHINFO_EXTENSION);
-				if (in_array(strtolower($extension), array('gif', 'jpeg', 'jpg', 'png'))) {
-					$name = (string)time() . '.' . $extension;
-					if (move_uploaded_file($tmp_name, IMAGES  . '..' . DS . 'files' . DS  . 'images' . DS . 'partners' . DS . $name) == true) {
-						unlink(IMAGES  . '..' . DS . 'files' . DS  . 'images' . DS . 'partners' . DS . $data['Partner']['partner_img']);
-						$d['Partner']['partner_img'] = $name;
-						if ($this->Partner->save($d, true, array('partner_type', 'partner_name', 'partner_desc', 'partner_img', 'partner_to_home', 'partner_to_page'))) {
-							$this->Session->setFlash("Le partenaire à bien été edité !", 'notif');
-							$this->redirect(array('controller' => 'partners', 'action' => 'index', 'admin' => true));
-						} else {
-							$this->Session->setFlash("Un problème est survenu !", 'notif', array('type' => 'error'));
-							$this->redirect^($this->referer());
-						}
+			$d['Partner']['partner_img'] = $this->upload_file($d['Partner']['partner_img'], 'partners');
+			if ($d['Partner']['partner_img'] != false) {
+				if ($this->delete_file($data['Partner']['partner_img'], 'partners')) {
+					if ($this->Partner->save($d, true, array('partner_type', 'partner_name', 'partner_desc', 'partner_img', 'partner_to_home', 'partner_to_page'))) {
+						$this->Session->setFlash("Le partenaire à bien été edité !", 'notif');
+						$this->redirect(array('controller' => 'partners', 'action' => 'index', 'admin' => true));
 					} else {
 						$this->Session->setFlash("Un problème est survenu !", 'notif', array('type' => 'error'));
-						$this->redirect($this->referer());
+						$this->redirect^($this->referer());
 					}
 				}
 			}
@@ -68,11 +49,11 @@ class PartnersController extends AppController {
 	public function admin_remove($id) {
 		$this->autoRender = false;
 		$partner = $this->Partner->find('first', array('conditions' => array('partner_id' => $id)));
-		$path = IMAGES  . '..' . DS . 'files' . DS  . 'images' . DS . 'partners' . DS . $partner['Partner']['partner_img'];
-		unlink($path);
-		$this->Partner->delete($id);
-		$this->Session->setFlash("Le partenaire a bien été supprimé !", 'notif');
-		$this->redirect($this->referer());
+		if ($this->delete_file($partner['Partner']['partner_img'], 'partners')) {
+			$this->Partner->delete($id);
+			$this->Session->setFlash("Le partenaire a bien été supprimé !", 'notif');
+			$this->redirect($this->referer());
+		}
 	}
 
 	public function admin_activated($type, $id) {
